Move test.py from tests to testing. Fixes #3106
