+++ networkx/__init__.py
@@ -121,7 +121,7 @@ from networkx.algorithms import *
 import networkx.linalg
 
 from networkx.linalg import *
-from networkx.tests.test import run as test
+from networkx.testing.test import run as test
 
 import networkx.drawing
 from networkx.drawing import *
