+++ networkx/testing/__init__.py
@@ -1,1 +1,2 @@
 from networkx.testing.utils import *
+from networkx.testing.test import run
